fix: modified some details in workforce
